backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
numpy
mercadopago
google-generativeai
chromadb
sentence-transformers
scikit-learn
openai
reportlab
cachetools
python-multipart
psycopg2-binary
requests
